net: increase the minimum size of the mbuf for TCP
Increase the minimum mbuf allocation size in the TCP to MCLBYTES (2048 bytes, before
it the minimum allocation was 256 bytes).

The consequent data buffers are compressed into the same mbuf in sbcompress() and therefore
this patch will, first, prevent the creation of enormous TCP mbuf clusters (256 elements for the
biggest LSO packet of 64K) and will keep them around the sane 32 elements.

Secondly, the bigger mbuf cluster elements let the HV work more efficiently when it needs
to copy them.

Signed-off-by: Vlad Zolotarov <vladz@cloudius-systems.com>