virtio: fix missed notification for HV without EVENT_IDX_CAP
This is a follow up patch for 053bcf8e18539d42719a2da0738ab7029a1f9a1c.

The similar race that has been fixed in the patch above may happen
when EVENT_IDX_CAP is not enabled by HV (e.g. GCE).

In that case the vring::kick() flow goes to the "else" which is not
covered by Avi's patch.

In the "else" we check the _used->_flags that is updated by the HV
(more specifically, we check for a VRING_USED_F_NO_NOTIFY bit in it) and
we use a "relaxed" memory order when we do it. Therefore this load operation
may be reordered before the store operation of _avail->_idx in vring::add_buf()
and thus we may erroneously decide not to wake a HV when we should have.

Fixed #379

Signed-off-by: Vlad Zolotarov <vladz@cloudius-systems.com>