Tx-no-IPI: initial commit
Change the per-CPU-Tx worker thread from beiing a single thread
that may be migrated by a load balancer to be a set of a per-CPU (pinned)
threads. This way an application thread when it needs to wake a worker it
will wake a worker on the same CPU and thus no IPIs are going to be sent.

 - Move the worker threads initialization into the per-CPU-Tx class (xmitter).
 - Move the initialization code from xmitter::poll_until() to xmitter
   constructor.
 - Name virtio::net tx worker threads as "virtio-tx-0,1,2..."
 - Name vmxnet3 tx worker threads as "vmxnet3-tx-0,1,2..."

Signed-off-by: Vlad Zolotarov <vladz@cloudius-systems.com>