virtio::net: Set infinite priority to worker threads
This patch utilizes the priority in order to ensure the fast response of a low latency thread.

In addition it prevents the context switch ping-pong for Rx and Tx worker threads when they
wake an application threads in the context of a netchannel and per-CPU-Tx flows correspondingly.

The above ping-pong may happen since the runtime of each worker is usually greater than a runtime of
a single application thread and scheduler will most likely switch to the application thread when it's
woken by a worker.

This patch:
   - Improves the performance of a ycsb workload-C cassandra test on a KVM Guest with 4 vCPU OSv
     Guest by 7.47%, 17.02%, 29.24% and 24.45% for 64, 128, 200 and 400 threads correspondingly.
   - Degrades the performance of the above test on single vCPU KVM Guest by ~10% for 64 and 128 threads
     and improves by 18.3% and 34.2% for 200 and 400 threads.
   - Degrades the redis-benchmark LRANGE_500 test result (50 sockets + 16 pilelines) on a KVM 2 vCPU
     Guest by 35%.

The Cassandra SMP case is the main beneficiary of this patch.  It's because it makes sure that the virtio threads -
running on one CPU - always get a chance to quickly distribute work to the other CPUs instead of letting
another thread on this single CPU run for a long time and cause the 3 other CPUs to become idle.

Signed-off-by: Vlad Zolotarov <vladz@cloudius-systems.com>

New in v3:
   - Added the explanation why Cassandra SMP case is the main beneficiary of this patch.