virtio::net: Set infinite priority to worker threads
This patch utilizes the priority in order to ensure the fast response of a low latency thread.

In addition it prevents the context switch ping-pong for Rx and Tx worker threads when they
wake an application threads in the context of a netchannel and per-CPU-Tx flows correspondingly.

The above ping-pong may happen since the runtime of each worker is usually greater than a runtime of
a single application thread and scheduler will most likely switch to the application thread when it's
woken by a worker.

This patch improves the performance of a ycsb workload-C cassandra test on 4 vCPU OSv
Guest by 7.47%, 17.02%, 29.24% and 24.45% for 64, 128, 200 and 400 threads correspondingly.

Signed-off-by: Vlad Zolotarov <vladz@cloudius-systems.com>