percpu_xmit: drop the tx-wroker-X priority to a default level
Drop the tx-wroker-X priority to a default level when we want Tx to migrate to a another CPU.
Another CPU's Tx worker will restore it when it aquires the RUNNING lock.

Without this patch the chance of other worker to actually aquire the RUNNING lock are
rather small. With this patch there is a high probability that the local tx-worker thread
will be preempted by another thread (since the trigger to migration is a long contineous running time)
and this will allow the remote tx-worker to get the RUNNING lock since it has a very high priority and
will most probably start running rather shortly after it's been woken up.

Signed-off-by: Vlad Zolotarov <vladz@cloudius-systems.com>