percpu_xmit: unpin the Tx worker
Pinned threads were hard to control in terms of load balancing -
leave it to a load balancer as it should be.

The main benefit of having a Tx-worker per-CPU was to minimize the number of
IPIs but after Gleb's patch (444a4e94e1) that minimizes IPIs globaly this patch hardly
improves it.

So, in order to simplify the code and to exclude the functionality (load balancing)
duplication this patch reverts the most of the (6589babdc7) patch.

Signed-off-by: Vlad Zolotarov <vladz@cloudius-systems.com>