sched: yield() don't overwrite the runtime of a yielding thread
Our yield() function guarantees that if there is another runnable thread, we
unconditionally switch to the other thread.
However, this was implemented by overwriting the yielding thread's "runtime" measurement
with a copy of the next thread's runtime - causing the scheduler to prefer that other
thread in the course of its normal flow.

Deleting the yield()ing thread's runtime history may create a lot of weird and unpredictable
situations, e.g. almost not running but yield()ing thread may not be getting a fair share of
the CPU's time when it would want to run because there is another frequently running thread on
on the same CPU; or otherwise, frequently running and yield()ing thread may "steal" the CPU time
in an unfair way from the less frequently running thread on the same CPU.

Namely calling yield() may cause our scheduler to become very unfair!

The solution is to change yield()'s implementation: Instead of changing the yielding thread's
runtime so that the scheduler's regular flow will take care of the yielding, we leave the runtime
unmodified and pass a "yielding" flag to the scheduler.

When the scheduler sees this flag, it behaves differently: It doesn't leave the current thread
running even if it still has the lowest runtime.

Also, when we switch to another thread, we need to pick some preemption time (so we'll eventually
return to the yielding thread). This time is passed to the scheduler as a parameter.

Note: This patch alone when applied to master HEAD 5c48ba8fa4567 introduce the regression of ~35%
      in redis LRANGE_500 (50 sockets, 16 pipelines) test on 2 vCPUs KVM Guest.

Signed-offdd-by: Vlad Zolotarov <vladz@cloudius-systems.com>

New in v3:
     - Removed added assert() in reschedule_from_interrupt().
     - Changed the type of "preempt_after": std::chrono::nanoseconds -> thread_runtime::duration.
     - Changed the default value of "preempt_after": 10us -> thread::thyst.
     - Changed the value of "preempt_after" virtio::net::txq::xmit_one_locked()
       calls thread::yield() with to 100us.
     - Changed the semantics of "preempt_after" parameter in reschedule_from_interrupt() a bit:
       when called from sched::yield() always set the preemption timer to this value.
     - Added some comments for sched::yield() and cpu::reschedule_from_interrupt().
     - Moved scheduler related constants declarations to sched.hh in order to be able to use
       sched::thyst as a default value for "preempt_after" parameter of sched::yield() and
       cpu::reschedule_from_interrupt().
     - Updated the patch description (thanks to Nadav for some ideas).